door/requirements.txt
raspberrypi-tm1637
RPi.GPIO
board
adafruit-circuitpython-sht31d

